# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tarn AI Evaluation API",
    "description": "AI suggests; the teacher decides. Every college sees only its own data.",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/v1/booklets/{booklet_id}/lock": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Open a booklet for review (take or refresh its lock)",
        "description": "The first open of a scored booklet starts its review (`in_review`). Call it again to keep the lock while reading (every write also refreshes it). Unapproved answers whose key, rubric, glossary or reference diagram changed since they were scored are sent for re-scoring (`rescoring`, `notices`).",
        "operationId": "open_booklet_api_v1_booklets__booklet_id__lock_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booklet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booklet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "423": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LockedOut"
                }
              }
            },
            "description": "Locked"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "review"
        ],
        "summary": "Close a booklet (release the caller's lock)",
        "description": "Nothing happens when the caller holds no lock on it.",
        "operationId": "close_booklet_api_v1_booklets__booklet_id__lock_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booklet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booklet Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/booklets/{booklet_id}/review": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "The review of a booklet: answers, suggestions, approvals, totals, sheets, lock",
        "operationId": "get_review_api_v1_booklets__booklet_id__review_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booklet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booklet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/booklets/{booklet_id}/result-sheets": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Every result sheet version of a booklet, oldest first",
        "operationId": "result_sheets_api_v1_booklets__booklet_id__result_sheets_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booklet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booklet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResultSheetOut"
                  },
                  "title": "Response Result Sheets Api V1 Booklets  Booklet Id  Result Sheets Get"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/booklets/{booklet_id}/answers/{answer_id}/approve": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Approve an answer: accept the AI's mark or override it",
        "description": "Both marks are kept. While the booklet is amended, approving the last draft issues the next result sheet version.",
        "operationId": "approve_answer_api_v1_booklets__booklet_id__answers__answer_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booklet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booklet Id"
            }
          },
          {
            "name": "answer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Answer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveAnswerIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "description": "Stale version (reload), a move the workflow does not allow, a new suggestion still on its way, or an approved answer that must be reopened first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "A mark, tag or text the rules refuse.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "423": {
            "description": "Another teacher has the booklet open, or the caller has not opened it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LockedOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/booklets/{booklet_id}/answers/{answer_id}/skip": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Skip an answer for now (approve it later)",
        "operationId": "skip_answer_api_v1_booklets__booklet_id__answers__answer_id__skip_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booklet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booklet Id"
            }
          },
          {
            "name": "answer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Answer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpectedVersionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "description": "Stale version (reload), a move the workflow does not allow, a new suggestion still on its way, or an approved answer that must be reopened first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "A mark, tag or text the rules refuse.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "423": {
            "description": "Another teacher has the booklet open, or the caller has not opened it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LockedOut"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorOut": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewOut": {
        "properties": {
          "booklet_id": {
            "type": "string",
            "format": "uuid",
            "title": "Booklet Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "uploaded",
              "processing",
              "needs_retake",
              "pages_ready",
              "reading",
              "text_ready",
              "segmented",
              "failed",
              "scored",
              "in_review",
              "approved",
              "amendment_in_progress",
              "approved_amended"
            ],
            "title": "Status"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "description": "Send it back as expected_version when approving the booklet or editing its text or segments."
          },
          "approved": {
            "type": "boolean",
            "title": "Approved",
            "description": "A result sheet stands (also during an amendment)."
          },
          "amendment_in_progress": {
            "type": "boolean",
            "title": "Amendment In Progress",
            "description": "The badge of an approved booklet."
          },
          "lock": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LockOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "can_approve": {
            "type": "boolean",
            "title": "Can Approve"
          },
          "waiting": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Waiting",
            "description": "Questions whose answers still need a decision."
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/ReviewAnswerOut"
            },
            "type": "array",
            "title": "Answers"
          },
          "totals": {
            "$ref": "#/components/schemas/TotalsOut",
            "description": "Teacher marks where approved, AI marks elsewhere, best N and OR applied."
          },
          "sheets": {
            "items": {
              "$ref": "#/components/schemas/ResultSheetOut"
            },
            "type": "array",
            "title": "Sheets",
            "description": "Every version issued, oldest first."
          },
          "rescoring": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Rescoring",
            "description": "On open: answers sent for re-scoring because their content changed."
          },
          "notices": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notices"
          }
        },
        "type": "object",
        "required": [
          "booklet_id",
          "status",
          "version",
          "approved",
          "amendment_in_progress",
          "lock",
          "can_approve",
          "waiting",
          "answers",
          "totals",
          "sheets"
        ],
        "title": "ReviewOut"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Access token from /api/v1/auth/login",
        "scheme": "bearer"
      }
    }
  }
}